refactor: split the cleaning test file into cleaning and loading_saving
